Several fixes to template query mechanism, to make its behavior match the manual, or to make the docs match the behavior, as apropriate.
Initial patches by Juergen M. F. Gleiss <juergen@gleiss.at>, with
corrections and expansions by Warren Young.


git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@1027 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ lib/query.h
@@ -689,7 +689,7 @@ private:
 	std::vector<std::string> parsed_names_;
 
 	/// \brief Maps template parameter names to their position value.
-	std::map<std::string, int> parsed_nums_;
+	std::map<std::string, short int> parsed_nums_;
 
 	/// \brief String buffer for storing assembled query
 	std::stringbuf sbuffer_;
